+++ CHANGELOG.md
@@ -11,10 +11,14 @@ Prefix your message with one of the following:
 - [Security] in case of vulnerabilities.
 -->
 
+## unreleased
+
+- [Changed] Link to single files with just the anchor.
+
 ## v0.0.1 - Jan 25, 2024
 
-- Remove unused dependencies.
-- Fix bug with SecureRandom.
+- [Fixed] Remove unused dependencies.
+- [Fixed] Fix bug with SecureRandom.
 
 ## v0.0.0 - Jan 25, 2024